+++ base/cvd/allocd/net/nftables.h
@@ -37,13 +37,13 @@ class Nftables {
                                    std::string_view table,
                                    std::string_view chain,
                                    std::string_view content) = 0;
-  virtual Result<uint32_t> AddRule(std::string_view family,
+  virtual Result<uint64_t> AddRule(std::string_view family,
                                    std::string_view table,
                                    std::string_view chain,
                                    std::string_view content) = 0;
   virtual Result<void> DeleteRule(std::string_view family,
                                   std::string_view table,
-                                  std::string_view chain, uint32_t handle) = 0;
+                                  std::string_view chain, uint64_t handle) = 0;
 };
 
 }  // namespace cuttlefish
